fix(trusty-review): correct inner_findings_count metric + fail-open smell (#1877)
Adds an authoritative `findings_count` field to ReviewResult, synced with
findings.len() at both canonical pipeline exit points (abort_dry and
finalize_review), so a downstream consumer can never again derive a
"findings count" by taking len() of the wrong (string) field.

Adds a `shallow_clean_review` telemetry flag + heuristic
(letter_grade::is_shallow_clean_review) that flags a zero-findings APPROVE
on a large diff whose output-token spend looks implausibly low for a
substantive pass (mirrors the reported pricerator#637 regression), and
caps the grade at B- in that case instead of defaulting to A+ on trust
alone. The genuinely fail-CLOSED paths (LLM/transport error, truncated
output, oversized diff) are untouched.

Splits models/mod.rs's inline test module into models/tests.rs to stay
under the 500-SLOC production cap after the new field documentation.
